Remove g_bans global
