+++ src/ban.h
@@ -13,29 +13,6 @@ struct BanInfo
 	time_t expiresAt;
 };
 
-struct ConnectBlock
-{
-	constexpr ConnectBlock(uint64_t lastAttempt, uint64_t blockTime, uint32_t count) :
-	    lastAttempt(lastAttempt), blockTime(blockTime), count(count)
-	{}
-
-	uint64_t lastAttempt;
-	uint64_t blockTime;
-	uint32_t count;
-};
-
-using IpConnectMap = std::map<Connection::Address, ConnectBlock>;
-
-class Ban
-{
-public:
-	bool acceptConnection(const Connection::Address& clientIP);
-
-private:
-	IpConnectMap ipConnectMap;
-	std::recursive_mutex lock;
-};
-
 class IOBan
 {
 public:
